Add proxy for Intern code coverage
Requests coming in for webapp JS files will be proxied to the Intern
proxy, assumed to be running on localhost:9000.

Uses UrlRewriteFilter to do the proxying - which requires the
following JARs:

- commons-codec-1.2.jar
- commons-httpclient-3.1.jar
- commons-logging-1.0.4.jar
- urlrewritefilter-4.0.3.jar

The web.xml and context.xml are now stored as templates, and copied and
modified into the webapp during build.

Use the "proxy" property during build, e.g.:

    ant -Dharviewer.folder=../harviewer2 -Dproxy=true